uniform network config
